refactor project page to server component
